requirements.txt
aioredis==1.3.1
appnope==0.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.1
bootstrap4==0.1.0
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
daphne==2.5.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.42.0
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-extensions==2.1.4
django-imagekit==4.0.2
django-map-widgets==0.3.0
django-polls @ file:///Users/ogawayasunori/mokuromi2/django-polls/dist/django-polls-0.1.tar.gz
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
fontawesome==5.10.1.post1
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.0
ipdb==0.11
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
more-itertools==8.4.0
msgpack==1.0.0
numpy==1.16.1
oauthlib==3.1.0
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.6
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.2.0
pytest-django==3.4.7
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
wcwidth==0.2.5
zope.interface==5.1.0
